added communication between the sensor and master
added communication between the sensor module and the master, where the sensor module sends the average of the last 10 measurements to the master via wifi. The value is send as a raw message, so the internal communication doesn't follow the http protocol.
The multithreaded code isn't super thread safe but so far I haven't encountered any bugs. The reason I didn't use _thread.lock is because this breaks the sensor reading and communication with the master.